project_dark_tracker/requirements.txt
Django==3.1.6
asgiref==3.3.1
djangorestframework==3.12.2
pip==19.0.3
pytz==2021.1
setuptools==40.8.0
sqlparse==0.4.1
django-cors-headers
pandas